Added ToolController class and moved weapon/tool functionality into it
