--- client/src/scenes/controllers/weapons/WeaponSettings.ts
@@ -1,4 +1,0 @@
-export default {
-  spear: {fallSpeedModifer: 1.5, stick: true, startVelModifer: 1, size: 1},
-  stone: {fallSpeedModifer: 3, stick: false, startVelModifer: .5, size: 1.5}
-} as {[name in Throwable]: {fallSpeedModifer: number, stick: boolean, startVelModifer: number, size: number}};
