+++ client/src/scenes/global.ts
@@ -12,6 +12,7 @@ export default {
   mobCount: 0 as number,
   inventory: {} as {[itemName: string]: number},
   Throwables: {'stone': true, 'spear': true} as {[type: string]: boolean},
+  Tools: {'stone': true, 'spear': true, 'bone_pickaxe': true} as {[type: string]: boolean},
   //This is for finding images only [for example stone is a tool and an env element] [this should be depending on where the image is located]
   EnvImages: ['stickyFern'] as EnvObj[],
   ItemImages: ['bone' , 'goo', 'stone', 'grapple', 'spear', 'bone_pickaxe'] as Drop[]
